requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
certifi==2025.11.12
environs==14.5.0
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
magic-filter==1.0.12
marshmallow==4.1.2
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
